OpenClass/requirements.txt
flask
flask-cors
openai
python-dotenv
mysql-connector-python
